Remove obsolete RemoteActivity
