+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,10 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity
-            android:name=".ui.RemoteActivity"
-            android:label="@string/title_activity_remote"
-            android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".ui.TestActivity" />
         <activity
             android:name=".ui.MainActivity"
+            android:theme="@style/AppTheme.NoActionBar"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
