+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,8 @@
 <resources>
-    <string name="app_name">LEDIRRemote</string>
+    <string name="app_name">IrRemote</string>
     <string name="action_settings">Settings</string>
     <string name="service_name_control">IRRemoteService</string>
     <string name="tile_name_power">LED Power</string>
     <string name="error_ir_not_available">No infrared emitter available. The application requires an IR blaster.</string>
     <string name="action_add_button">Add button</string>
-    <string name="title_activity_remote">RemoteActivity</string>
 </resources>
